refactor: Move even even more tests into their own files
With this, all the tests of the "big" files are in their own files.
